requirements.txt
minify-html==0.15.5
newsapi-python==0.2.7
serpapi==1.3.0
google-search-results==2.4.2
css_inline==0.12.0
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4